+++ Sources/RichTextKit/_Deprecated/RichTextStyle+Deprecated.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "RichTextStyle.Button")
+public typealias RichTextStyleButton = RichTextStyle.Button
+
+@available(*, deprecated, renamed: "RichTextStyle.Toggle")
+public typealias RichTextStyleToggle = RichTextStyle.Toggle
+
+@available(*, deprecated, renamed: "RichTextStyle.ToggleGroup")
+public typealias RichTextStyleToggleGroup = RichTextStyle.ToggleGroup
+
+@available(*, deprecated, renamed: "RichTextStyle.ToggleStack")
+public typealias RichTextStyleToggleStack = RichTextStyle.ToggleStack